feat: Rename the custom event of graphviz-script-editor from "input" to "update"
Let it be easily differentiated from the `input` event just by the event type.

BREAKING CHANGE: If you were listening to the `input` event, change the event name
to `update`. The event `detail` didn't change - it still contains the updated script.
Why did I rename the event? The original `input` event is triggered as composed
and bubbled out of the shadow DOM. The `input` event handler caught it. The custom
`input` event just followed it, which made its processing more difficult - for example,
by checking the class of the event to be `CustomEvent`. Using a different event type
- `update` - will avoid duplicate event processing without any extra conditions.